added pad left fnc
