+++ src/strategies/index.ts
@@ -9,7 +9,7 @@ const strategies = {
 };
 
 export function getStrategy(address: string): Strategy | null {
-  const strategy = strategies[utils.encoding.hexPadRight(address)];
+  const strategy = strategies[utils.encoding.hexPadLeft(address)];
   if (!strategy) return null;
 
   return strategy;
