+++ src/utils/encoding/eth-sig.ts
@@ -1,18 +1,5 @@
 import { SplitUint256 } from '../split-uint256';
 
-export function hexPadRight(s: string) {
-  // Remove prefix
-  if (s.startsWith('0x')) {
-    s = s.substring(2);
-  }
-  // Odd length, need to prefix with a 0
-  if (s.length % 2 != 0) {
-    s = `0${s}`;
-  }
-  const numZeroes = 64 - s.length;
-  return `0x${s + '0'.repeat(numZeroes)}`;
-}
-
 // Extracts and returns the `r, s, v` values from a `signature`
 export function getRSVFromSig(sig: string) {
   if (sig.startsWith('0x')) {
