+++ src/utils/encoding/index.ts
@@ -5,3 +5,4 @@ export * from './starknet-commit';
 export * from './starknet-execution-params';
 export * from './eth-sig';
 export * from './storage-vars';
+export * from './padding';
